Safari-609.1.11.1/LayoutTests/fast/forms/watchos/delete-content-in-text-field-expected.txt

PASS inputEventAfterInsertingText.type is 'input'
PASS inputEventAfterInsertingText.inputType is 'insertText'
PASS inputEventAfterInsertingText.data is 'testing'
PASS valueAfterInsertingText is 'testing'
PASS inputEventAfterDeletingText.type is 'input'
PASS inputEventAfterDeletingText.inputType is 'deleteContent'
PASS inputEventAfterDeletingText.data is null
PASS valueAfterDeletingText is ''
PASS successfullyParsed is true

TEST COMPLETE
